requirements.txt
python-telegram-bot==22.3
APScheduler==3.11.0
python-dotenv==1.1.1
requests==2.32.4
httpx==0.28.1
httpcore==1.0.9
anyio==4.10.0
h11==0.16.0
exceptiongroup==1.3.0
sniffio==1.3.1
typing_extensions==4.14.1
tzlocal==5.3.1
